Refactor: Extract UncaughtExceptionBase and some fixes in documentation
