+++ src/main/java/ir/sahab/uncaughtexceptionrule/UncaughtExceptionRule.java
@@ -17,7 +17,7 @@ import org.junit.runners.model.Statement;
  * {@literal @}Rule
  *  UncaughtExceptionRule rule = new UncaughtExceptionRule();
  * </pre>
- *
+ * <p>
  * By default test will be failed if exception occurs. You can also override default behavior for a
  * specific test case. Where you expect an exception in a working thread (but not others), put the
  * line below at the bottom of that specific test case:
@@ -26,56 +26,29 @@ import org.junit.runners.model.Statement;
  *  Assert.assertTrue(rule.getException() instanceof ...);
  *  rule.clearException();  // Clears the exception to avoid failure on tear down.
  * </pre>
- *
+ * <p>
  * Note that the cached exception from a previous test will be cleared before running new tests and
  * just fails that test and is no need to manually clear exception.
  */
 
-public class UncaughtExceptionRule implements TestRule {
-
-    private Throwable unhandledException = null;
-    private Thread.UncaughtExceptionHandler oldHandler;
+public class UncaughtExceptionRule extends UncaughtExceptionBase implements TestRule {
 
     @Override
     public Statement apply(Statement base, Description description) {
-        oldHandler = Thread.getDefaultUncaughtExceptionHandler();
-
         return new Statement() {
             @Override
             public void evaluate() throws Throwable {
-                unhandledException = null;
-                // Set default exception handler so that we can catch exception on all working
-                // threads.
-                Thread.setDefaultUncaughtExceptionHandler((thread, throwable) -> {
-                    unhandledException = throwable;
-                    // Let whatever that was going to happen to this exception, still happen.
-                    if (oldHandler != null) {
-                        oldHandler.uncaughtException(thread, throwable);
-                    }
-                });
-
+                registerHandler();
                 try {
                     // Run one test or all tests of a test class depending on whether it is a
                     // @Rule or @ClassRule
                     base.evaluate();
                 } finally {
-                    // Throw the exception in the main thread of the test. So it cause test failure.
-                    if (unhandledException != null) {
-                        throw unhandledException;
-                    }
-                    // Set default exception handler to the old one.
-                    Thread.setDefaultUncaughtExceptionHandler(oldHandler);
+                    afterHandle();
                 }
             }
         };
     }
 
-    public Throwable getException() {
-        return unhandledException;
-    }
-
-    public void clearException() {
-        this.unhandledException = null;
-    }
 }
